23.03.31 Hospital Allocation Datalink - Yilun Li

+++ testDisasterEvent/app/src/main/java/com/example/testdisasterevent/data/model/DisasterDetail.java
@@ -4,14 +4,14 @@ public class DisasterDetail {
     private int radius;
     private String location;
     private String happenTime;
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
     private String disasterType;
     private String GardaUID;
     private int isUpdate;
 
 
-    public DisasterDetail(int radius, String location, String happenTime, float latitude, float longitude, String disasterType, String GardaUID, int isUpdate) {
+    public DisasterDetail(int radius, String location, String happenTime, double latitude, double longitude, String disasterType, String GardaUID, int isUpdate) {
         this.radius = radius;
         this.location = location;
         this.happenTime = happenTime;
@@ -46,19 +46,19 @@ public class DisasterDetail {
         this.happenTime = happenTime;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
